Create the computerPlay function
NOTE: The game is going to run purely on the browser console for now so
all code will be written in game.js while index.html will remain a blank
HTML skeleton.

The game is going to be played against a computer. This function called
computerPlay randomly returns either 'Rock', 'Paper' or 'Scissors'.
We'll use this function in the game to make the computer's play.