Create the game function
The game function calls the playRound function 5 times (simulating 5
rounds). After the 5 rounds, a winner is declared. Ties do not count as
completed rounds (in other words, we do not proceed to the next round
until either the player or the computer win the round). All necessary
output is logged to the console and the user is prompted for their
selection of 'Rock', 'Paper', or 'Scissors' each round.